+++ src/webComponents/components.ts
@@ -3,6 +3,7 @@ import FormField from "./form/FormField";
 import ImageUploader from "./form/ImageUploader";
 import PrettyButton from "./form/PrettyButton";
 import RadioSelection from "./form/RadioSelection";
+import RangeInput from "./form/RangeInput";
 import TextInput from "./form/TextInput";
 import GameControl from "./gameplay/GameControl";
 import GameGauge from "./gameplay/GameGauge";
@@ -40,6 +41,7 @@ const components: ComponentRegistration[] = [
   { tag: "form-field", Constructor: FormField },
   { tag: "social-link", Constructor: SocialLink },
   { tag: "text-link", Constructor: TextLink },
+  { tag: "range-input", Constructor: RangeInput },
 ];
 
 const registerComponentWithDependencies = ({