+++ changelog.md
@@ -14,6 +14,7 @@
   - Bugfix: Game no longer resumes if login dialog is open
 - Settings dialog:
   - Date of Birth updated to optional
+  - Introduce audio volume slider
 
 # 0.8.7 Beta - 02/02/2026 13:30 UTC+1